+++ safileio.c
@@ -124,7 +124,7 @@ void SASetupDefaultHooks(SAHooks *psHooks)
 
 static wchar_t *Utf8ToWideChar(const char *pszFilename)
 {
-    const int nMulti = strlen(pszFilename) + 1;
+    const int nMulti = (int)strlen(pszFilename) + 1;
     const int nWide =
         MultiByteToWideChar(CP_UTF8, 0, pszFilename, nMulti, 0, 0);
     if (nWide == 0)
